Rename functions.signature module
Mypy was complaining that the name 'signature' was ambiguous in the
vyper/functions/__init__.py file because there was a module
(signature.py) defined at that level as well as a symbol 'signature'
being imported from vyper.functions.functions.
